web_demo/chroma_server/requirements.txt
chromadb>=0.5.0
sentence-transformers>=2.2.2
underthesea>=6.8.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dotenv>=1.0.0
tqdm>=4.66.1
numpy>=1.26.0
flask>=2.0.0
requests>=2.31.0